Fixes.
- GuiServer - Allow for packets to overflow MAX_PACKET_SIZE while building as long as the resulting packet size after compression fits on uint16.
- GUIServer - Remove individual packet header size, this is now always handled when compressing packets.

+++ SniffBrowser/Network/NetworkClient.cs
@@ -102,7 +102,7 @@ namespace SniffBrowser.Network
                 {
                     // Header.
                     case sizeof(uint):
-                        packetSize = BitConverter.ToUInt32(headerBuffer, 0);
+                        packetSize = BitConverter.ToUInt16(headerBuffer, 0);
 
                         if (packetSize == 0)
                         {
